+++ example/lib/boundaries.dart
@@ -1,14 +1,15 @@
 import 'dart:ui';
 
+import 'package:flame_forge2d/flame_forge2d.dart';
+import 'package:flame_forge2d/forge2d_game.dart';
 import 'package:forge2d/forge2d.dart';
 import 'package:flame/palette.dart';
+import 'package:flame/game.dart';
 import 'package:flame_forge2d/body_component.dart';
-import 'package:flame_forge2d/viewport.dart';
 
-List<Wall> createBoundaries(Viewport viewport) {
-  final Vector2 screenSize = viewport.size / viewport.scale;
-  final Vector2 topLeft = (screenSize / 2) * -1;
-  final Vector2 bottomRight = screenSize / 2;
+List<Wall> createBoundaries(Forge2DGame game) {
+  final Vector2 topLeft = Vector2.zero();
+  final Vector2 bottomRight = game.screenToWorld(game.viewport.effectiveSize);
   final Vector2 topRight = Vector2(bottomRight.x, topLeft.y);
   final Vector2 bottomLeft = Vector2(topLeft.x, bottomRight.y);
 
@@ -21,33 +22,24 @@ List<Wall> createBoundaries(Viewport viewport) {
 }
 
 class Wall extends BodyComponent {
-  Paint paint = BasicPalette.white.paint;
+  Paint paint = BasicPalette.white.paint();
   final Vector2 start;
   final Vector2 end;
 
   Wall(this.start, this.end);
 
   @override
-  void renderPolygon(Canvas canvas, List<Offset> coordinates) {
-    final start = coordinates[0];
-    final end = coordinates[1];
-    canvas.drawLine(start, end, paint);
-  }
-
-  @override
   Body createBody() {
-    final PolygonShape shape = PolygonShape();
-    shape.setAsEdge(start, end);
+    final shape = EdgeShape()..set(start, end);
 
-    final fixtureDef = FixtureDef()
-      ..shape = shape
+    final fixtureDef = FixtureDef(shape)
       ..restitution = 0.0
-      ..friction = 0.1;
+      ..friction = 0.3;
 
     final bodyDef = BodyDef()
       ..userData = this // To be able to determine object in collision
       ..position = Vector2.zero()
-      ..type = BodyType.STATIC;
+      ..type = BodyType.static;
 
     return world.createBody(bodyDef)..createFixture(fixtureDef);
   }