+++ example/lib/mouse_joint_sample.dart
@@ -1,7 +1,6 @@
 import 'package:flame_forge2d/flame_forge2d.dart';
 import 'package:flame_forge2d/forge2d_game.dart';
 import 'package:flame/gestures.dart';
-import 'package:flutter/material.dart';
 import 'package:flame/extensions.dart';
 
 import 'balls.dart';
@@ -9,54 +8,49 @@ import 'boundaries.dart';
 import 'circle_stress_sample.dart';
 
 class MouseJointSample extends Forge2DGame with MultiTouchDragDetector {
-  @override
-  bool debugMode = true;
+  late Ball ball;
+  late Body groundBody;
+  MouseJoint? mouseJoint;
 
-  Ball ball;
-  Body groundBody;
-  MouseJoint mouseJoint;
-
-  MouseJointSample()
-      : super(
-          scale: 8.0,
-          gravity: Vector2(0, -10.0),
-        );
+  MouseJointSample() : super(gravity: Vector2(0, -10.0));
 
   @override
   Future<void> onLoad() async {
-    final boundaries = createBoundaries(viewport);
+    await super.onLoad();
+    final boundaries = createBoundaries(this);
     boundaries.forEach(add);
 
+    final center = screenToWorld(viewport.effectiveSize / 2);
     groundBody = world.createBody(BodyDef());
-    ball = Ball(viewport.getWorldToScreen(Vector2(0, 0)), radius: 5);
+    ball = Ball(center, radius: 5);
     add(ball);
-    add(CornerRamp());
-    add(CornerRamp(isMirrored: true));
+    add(CornerRamp(center));
+    add(CornerRamp(center, isMirrored: true));
   }
 
   @override
-  void onReceiveDrag(DragEvent drag) {
-    drag.onUpdate = (DragUpdateDetails details) {
-      MouseJointDef mouseJointDef = MouseJointDef()
-        ..maxForce = 3000 * ball.body.mass * 10
-        ..dampingRatio = 0.1
-        ..frequencyHz = 5
-        ..target.setFrom(ball.body.position)
-        ..collideConnected = false
-        ..bodyA = groundBody
-        ..bodyB = ball.body;
+  bool onDragUpdate(int pointerId, DragUpdateInfo details) {
+    MouseJointDef mouseJointDef = MouseJointDef()
+      ..maxForce = 3000 * ball.body.mass * 10
+      ..dampingRatio = 0.1
+      ..frequencyHz = 5
+      ..target.setFrom(ball.body.position)
+      ..collideConnected = false
+      ..bodyA = groundBody
+      ..bodyB = ball.body;
 
-      mouseJoint ??= world.createJoint(mouseJointDef);
+    mouseJoint ??= world.createJoint(mouseJointDef) as MouseJoint;
 
-      mouseJoint.setTarget(
-          viewport.getScreenToWorld(details.globalPosition.toVector2()));
-    };
+    mouseJoint?.setTarget(details.eventPosition.game);
+    return false;
+  }
 
-    drag.onEnd = (DragEndDetails details) {
-      world.destroyJoint(mouseJoint);
-      mouseJoint = null;
-    };
-
-    super.onReceiveDrag(drag);
+  bool onDragEnd(int pointerId, DragEndInfo details) {
+    if (mouseJoint == null) {
+      return true;
+    }
+    world.destroyJoint(mouseJoint!);
+    mouseJoint = null;
+    return false;
   }
 }