+++ example/lib/position_body_sample.dart
@@ -5,41 +5,46 @@ import 'package:forge2d/forge2d.dart';
 import 'package:flame/components.dart';
 import 'package:flame/gestures.dart';
 import 'package:flame_forge2d/forge2d_game.dart';
-import 'package:flutter/material.dart' hide Image;
 
 import 'boundaries.dart';
 
 class ChopperBody extends PositionBodyComponent {
-  ChopperBody(PositionComponent component) : super(component, component.size);
+  final Vector2 position;
+
+  ChopperBody(
+    this.position,
+    PositionComponent component,
+  ) : super(component, component.size);
 
   @override
   Body createBody() {
     final shape = CircleShape()..radius = size.x / 4;
-    final fixtureDef = FixtureDef();
-    fixtureDef.userData = this; // To be able to determine object in collision
-    fixtureDef.shape = shape;
-    fixtureDef.restitution = 0.3;
-    fixtureDef.density = 1.0;
-    fixtureDef.friction = 0.2;
+    final fixtureDef = FixtureDef(shape)
+      ..userData = this // To be able to determine object in collision
+      ..restitution = 0.8
+      ..density = 1.0
+      ..friction = 0.2;
 
+    final velocity = (Vector2.random() - Vector2.random()) * 200;
     final bodyDef = BodyDef()
-      ..position = viewport.getScreenToWorld(positionComponent.position)
-      ..angle = positionComponent.x / 2 * 3.14
-      ..linearVelocity = (Vector2.all(0.5) - Vector2.random()) * 200
-      ..type = BodyType.DYNAMIC;
+      ..position = position
+      ..angle = velocity.angleTo(Vector2(1, 0))
+      ..linearVelocity = velocity
+      ..type = BodyType.dynamic;
     return world.createBody(bodyDef)..createFixture(fixtureDef);
   }
 }
 
 class PositionBodySample extends Forge2DGame with TapDetector {
-  Image chopper;
-  SpriteAnimation animation;
+  late Image chopper;
+  late SpriteAnimation animation;
 
-  PositionBodySample() : super(scale: 10.0, gravity: Vector2.zero());
+  PositionBodySample() : super(gravity: Vector2.zero());
 
   @override
   Future<void> onLoad() async {
-    chopper = await images.load('animations/chopper.png');
+    await super.onLoad();
+    chopper = await images.load('chopper.png');
 
     animation = SpriteAnimation.fromFrameData(
       chopper,
@@ -51,19 +56,19 @@ class PositionBodySample extends Forge2DGame with TapDetector {
       ),
     );
 
-    final boundaries = createBoundaries(viewport);
+    final boundaries = createBoundaries(this);
     boundaries.forEach(add);
   }
 
   @override
-  void onTapDown(TapDownDetails details) {
+  void onTapDown(TapDownInfo details) {
     super.onTapDown(details);
-    final Vector2 position =
-        Vector2(details.localPosition.dx, details.localPosition.dy);
+    final Vector2 position = details.eventPosition.game;
     final spriteSize = Vector2.all(10);
-    final animationComponent =
-        SpriteAnimationComponent.fromSpriteAnimation(spriteSize, animation);
-    animationComponent.position = position;
-    add(ChopperBody(animationComponent));
+    final animationComponent = SpriteAnimationComponent(
+      animation: animation,
+      size: spriteSize,
+    );
+    add(ChopperBody(position, animationComponent));
   }
 }