+++ example/lib/sprite_body_sample.dart
@@ -5,7 +5,6 @@ import 'package:flame/components.dart';
 import 'package:flame/gestures.dart';
 import 'package:flame_forge2d/forge2d_game.dart';
 import 'package:flame_forge2d/sprite_body_component.dart';
-import 'package:flutter/material.dart' hide Image;
 
 import 'boundaries.dart';
 
@@ -18,44 +17,42 @@ class Pizza extends SpriteBodyComponent {
   Body createBody() {
     final PolygonShape shape = PolygonShape();
 
-    final v1 = Vector2(0, size.y / 2);
-    final v2 = Vector2(size.x / 2, -size.y / 2);
-    final v3 = Vector2(-size.x / 2, -size.y / 2);
-    final vertices = [v1, v2, v3];
-    shape.set(vertices, vertices.length);
+    final vertices = [
+      Vector2(-size.x / 2, -size.y / 2),
+      Vector2(size.x / 2, -size.y / 2),
+      Vector2(0, size.y / 2),
+    ];
+    shape.set(vertices);
 
-    final fixtureDef = FixtureDef();
-    fixtureDef.userData = this; // To be able to determine object in collision
-    fixtureDef.shape = shape;
-    fixtureDef.restitution = 0.3;
-    fixtureDef.density = 1.0;
-    fixtureDef.friction = 0.2;
+    final fixtureDef = FixtureDef(shape)
+      ..userData = this // To be able to determine object in collision
+      ..restitution = 0.3
+      ..density = 1.0
+      ..friction = 0.2;
 
-    final bodyDef = BodyDef();
-    bodyDef.position = viewport.getScreenToWorld(_position);
-    bodyDef.angle = (_position.x + _position.y) / 2 * 3.14;
-    bodyDef.type = BodyType.DYNAMIC;
+    final bodyDef = BodyDef()
+      ..position = _position
+      ..angle = (_position.x + _position.y) / 2 * 3.14
+      ..type = BodyType.dynamic;
     return world.createBody(bodyDef)..createFixture(fixtureDef);
   }
 }
 
 class SpriteBodySample extends Forge2DGame with TapDetector {
-  Image _pizzaImage;
+  late Image _pizzaImage;
 
-  SpriteBodySample() : super(scale: 4.0, gravity: Vector2(0, -10.0));
+  SpriteBodySample() : super(gravity: Vector2(0, -10.0));
 
   @override
   Future<void> onLoad() async {
     _pizzaImage = await images.load('pizza.png');
-    final boundaries = createBoundaries(viewport);
-    boundaries.forEach(add);
+    addAll(createBoundaries(this));
   }
 
   @override
-  void onTapDown(TapDownDetails details) {
+  void onTapDown(TapDownInfo details) {
     super.onTapDown(details);
-    final Vector2 position =
-        Vector2(details.localPosition.dx, details.localPosition.dy);
+    final Vector2 position = details.eventPosition.game;
     add(Pizza(position, _pizzaImage));
   }
 }