+++ lib/body_component.dart
@@ -1,64 +1,94 @@
 import 'dart:ui';
 
 import 'package:flame/components.dart';
+import 'package:flame/extensions.dart';
+import 'package:flame/game.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:forge2d/forge2d.dart' hide Timer, Vector2;
 
 import 'forge2d_game.dart';
-import 'viewport.dart';
 
 /// Since a pure BodyComponent doesn't have anything drawn on top of it,
-/// it is a good idea to turn on debudMode for it so that the bodies can be seen
-
+/// it is a good idea to turn on [debugMode] for it so that the bodies can be
+/// seen
 abstract class BodyComponent<T extends Forge2DGame> extends BaseComponent
     with HasGameRef<T> {
-  static const maxPolygonVertices = 8;
   static const defaultColor = const Color.fromARGB(255, 255, 255, 255);
+  late Body body;
+  late Paint paint;
 
-  Body body;
-  Paint paint;
+  /// [debugMode] is true by default for body component since otherwise
+  /// nothing is rendered for it, if you render something on top of the
+  /// [BodyComponent], or doesn't want it to be seen, just set it to false.
+  /// [SpriteBodyComponent] and [PositionBodyComponent] has it set to false by
+  /// default.
+  @override
+  bool debugMode = true;
 
-  BodyComponent({this.paint}) {
-    paint ??= Paint()..color = defaultColor;
+  BodyComponent({Paint? paint}) {
+    this.paint = paint ?? (Paint()..color = defaultColor);
   }
 
-  /// You should create the [Body] in this method when you extend
+  /// You should create the Forge2D [Body] in this method when you extend
   /// the BodyComponent
   Body createBody();
 
+  @mustCallSuper
+  @override
+  Future<void> onLoad() async {
+    body = createBody();
+  }
+
   World get world => gameRef.world;
-  Viewport get viewport => gameRef.viewport;
+  Camera get camera => gameRef.camera;
+  Vector2 get center => body.worldCenter;
+  double get angle => body.angle;
+
+  /// The matrix used for preparing the canvas
+  final Matrix4 _transform = Matrix4.identity();
+  double? _lastAngle;
+
+  @override
+  void prepareCanvas(Canvas canvas) {
+    if (_transform.m14 != body.position.x ||
+        _transform.m24 != body.position.y ||
+        _lastAngle != angle) {
+      _transform.setIdentity();
+      _transform.scale(1, -1);
+      _transform.translate2(body.position);
+      _transform.rotateZ(angle);
+      _lastAngle = angle;
+    }
+    canvas.transform(_transform.storage);
+  }
 
   @override
   void renderDebugMode(Canvas canvas) {
     for (Fixture fixture in body.fixtures) {
-      switch (fixture.getType()) {
-        case ShapeType.CHAIN:
+      switch (fixture.type) {
+        case ShapeType.chain:
           _renderChain(canvas, fixture);
           break;
-        case ShapeType.CIRCLE:
+        case ShapeType.circle:
           _renderCircle(canvas, fixture);
           break;
-        case ShapeType.EDGE:
+        case ShapeType.edge:
           _renderEdge(canvas, fixture);
           break;
-        case ShapeType.POLYGON:
+        case ShapeType.polygon:
           _renderPolygon(canvas, fixture);
           break;
       }
     }
   }
 
-  Vector2 get center => body.worldCenter;
-
   void _renderChain(Canvas canvas, Fixture fixture) {
-    assert(viewport != null, "Needs the viewport set to be able to render");
-    final ChainShape chainShape = fixture.shape;
-    final List<Offset> points = [];
-    for (int i = 0; i < chainShape.vertexCount; i++) {
-      points.add(_vertexToScreen(chainShape.getVertex(i)));
-    }
-    renderChain(canvas, points);
+    final ChainShape chainShape = fixture.shape as ChainShape;
+    renderChain(
+      canvas,
+      chainShape.vertices.map((v) => v.toOffset()).toList(growable: false),
+    );
   }
 
   void renderChain(Canvas canvas, List<Offset> points) {
@@ -67,10 +97,8 @@ abstract class BodyComponent<T extends Forge2DGame> extends BaseComponent
   }
 
   void _renderCircle(Canvas canvas, Fixture fixture) {
-    assert(viewport != null, "Needs the viewport set to be able to render");
-    final CircleShape circle = fixture.shape;
-    final center = _vertexToScreen(circle.position);
-    renderCircle(canvas, center, circle.radius * viewport.scale);
+    final CircleShape circle = fixture.shape as CircleShape;
+    renderCircle(canvas, circle.position.toOffset(), circle.radius);
   }
 
   void renderCircle(Canvas canvas, Offset center, double radius) {
@@ -78,47 +106,29 @@ abstract class BodyComponent<T extends Forge2DGame> extends BaseComponent
   }
 
   void _renderPolygon(Canvas canvas, Fixture fixture) {
-    assert(viewport != null, "Needs the viewport set to be able to render");
-    final PolygonShape polygon = fixture.shape;
-    assert(polygon.count <= maxPolygonVertices);
-
-    final List<Offset> points = [];
-    for (int i = 0; i < polygon.count; i++) {
-      points.add(_vertexToScreen(polygon.vertices[i]));
-    }
-
-    renderPolygon(canvas, points);
+    final PolygonShape polygon = fixture.shape as PolygonShape;
+    renderPolygon(
+      canvas,
+      polygon.vertices.map((v) => v.toOffset()).toList(growable: false),
+    );
   }
 
   void renderPolygon(Canvas canvas, List<Offset> points) {
     final path = Path()..addPolygon(points, true);
-
     canvas.drawPath(path, paint);
   }
 
   void _renderEdge(Canvas canvas, Fixture fixture) {
     final edge = fixture.shape as EdgeShape;
-    final p1 = _vertexToScreen(edge.vertex1);
-    final p2 = _vertexToScreen(edge.vertex2);
-    renderEdge(canvas, p1, p2);
+    renderEdge(canvas, edge.vertex1.toOffset(), edge.vertex2.toOffset());
   }
 
   void renderEdge(Canvas canvas, Offset p1, Offset p2) {
     canvas.drawLine(p1, p2, paint);
   }
 
-  Offset _vertexToScreen(Vector2 vertex) {
-    return viewport.getWorldToScreen(body.getWorldPoint(vertex)).toOffset();
-  }
-
   @override
-  bool checkOverlap(Vector2 point) {
-    final worldPoint = viewport.getScreenToWorld(point);
-    for (Fixture fixture in body.fixtures) {
-      if (fixture.testPoint(worldPoint)) {
-        return true;
-      }
-    }
-    return false;
+  bool containsPoint(Vector2 point) {
+    return body.fixtures.any((fixture) => fixture.testPoint(point));
   }
 }